cleaned up structure, rm v1 app.py
