+++ imageSplitCatalog/app.py
@@ -1,13 +1,11 @@
 import os
 import botocore
 import boto3
+from tifffile import tifffile
 from datetime import datetime
 from PIL import Image
 from PIL.ExifTags import TAGS, GPSTAGS
 import numpy as np
-import picamraw
-from tifffile import tifffile
-from picamraw import PiRawBayer, PiCameraVersion
 
 
 class Envs:
@@ -83,7 +81,7 @@ def get_coordinates(geotags):
 
 
 def lambda_handler(event, context):
-    print('## EVENT')
+    print('#### START - Event Received ####')
     print(event)
     s3_key = event["Records"][0]["s3"]["object"]["key"]
     s3_bucket = event["Records"][0]["s3"]["bucket"]["name"]
@@ -95,7 +93,7 @@ def lambda_handler(event, context):
     source_uri = (f"s3://{s3_bucket}/{s3_key}")
     tiff_uri = (f"s3://{Envs.tiff_bucket}/{source_camera}/{tiff_filename}")
     jpeg_uri = (f"s3://{Envs.jpeg_bucket}/{source_camera}/{jpeg_filename}")
-    print('## ENV')
+    print('## ENV VALUES ##')
     print(f"Source Bucket: {s3_bucket}")
     print(f"Source Key: {s3_key}")
     print(f"Image Name: {image_name}")
@@ -107,33 +105,50 @@ def lambda_handler(event, context):
     print(f"JPEG Location: {jpeg_uri}")
     print(f"DynamoDB Table: {Envs.ddb_table}")
 
+    print(f"## Image Processing ##")
     s3.download_file(s3_bucket, s3_key, source_image_tmp)
     # Split off the TIFF
-    source_image = picamraw.PiRawBayer(
-        filepath=source_image_tmp,  camera_version=picamraw.PiCameraVersion.V1, sensor_mode=0)
-    c = source_image.to_rgb()
-    #! Keep the RGB as the split tiff
+    source_image = open(source_image_tmp, "rb")
+    rawFileOffset = 18711040
+    rawHeaderSize = 32768
+    rawDataOffset = rawFileOffset - rawHeaderSize
+    RELATIVETOEND = 2
 
-    # c = source_image.bayer_array   # A 16-bit 2D numpy array of the bayer data
-    # c = source_image.bayer_order   # A `BayerOrder` enum that describes the arrangement of the R,G,G,B pixels in the bayer_array
-    # c = source_image.to_rgb()      # A 16-bit 3D numpy array of bayer data collapsed into RGB channels (see docstring for details).
-    # c = source_image.to_3d()       # A 16-bit 3D numpy array of bayer data split into RGB channels (see docstring for details).
+    source_image.seek(-rawDataOffset,RELATIVETOEND)
+    pixelsPerLine=4056
+    pixelsPerColumn=3040
+    usedLineBytes=int(pixelsPerLine*12/8)# source_image = picamraw.PiRawBayer(
+    s=np.zeros((pixelsPerLine,pixelsPerColumn),dtype='uint16')
+    # do this at least 3040 times though the precise number of lines is a bit unclear
+    for i in range(pixelsPerColumn):
+        # read in 6112 bytes, but only 6084 will be used
+        bdLine = source_image.read(6112)
+        gg=np.array(list(bdLine[0:usedLineBytes]),dtype='uint16')
+        s[0::2,i] = (gg[0::3]<<4) + (gg[2::3]&15)
+        s[1::2,i] = (gg[1::3]<<4) + (gg[2::3]>>4)
+    # form superpixel.tiff-pixel array
+    sp=np.zeros((int(pixelsPerLine/2),int(pixelsPerColumn/2),3),dtype='uint16')
+    sp[:,:,0]=s[1::2,1::2]                      # red
+    sp[:,:,1]=(s[0::2,1::2]+s[1::2,0::2])/2     # green
+    sp[:,:,2]=s[0::2,0::2]                      # blue
+    sp = np.multiply(sp,16) # adjusting the image to be saturated correctly(it was imported from 12bit into a 16bit) so it is a factor of 16 dimmer than should be, i.e this conversion
+    sp=sp.transpose()
+    array = ((sp+0.5).astype('uint16'))
 
-    array = (64*c.astype(np.uint16))
     tifffile.imwrite((f"/tmp/{tiff_filename}"), array, photometric='rgb')
-
     s3.upload_file((f"/tmp/{tiff_filename}"), Envs.tiff_bucket,
                    (f"{source_camera}/{tiff_filename}"))
-
     print('TIFF Uploaded')
-    # Splitting off the JPEG, saving and applying EXIF from source to JPEG
+
+    # Split off the JPEG, save and apply EXIF from source to JPEG
     jpeg = Image.open(source_image_tmp).convert("RGB")
     exif = jpeg.info['exif']
     print('EXIF saved on JPEG')
     jpeg.save((f"/tmp/{jpeg_filename}"), 'JPEG', exif=exif)
-    print('JPEG Uploaded')
     s3.upload_file((f"/tmp/{jpeg_filename}"), Envs.jpeg_bucket,
                    (f"{source_camera}/{jpeg_filename}"))
+    print('JPEG Uploaded')
+
     # Return Exif tags
     exif = get_exif(source_image_tmp)
     if not exif:
@@ -151,22 +166,31 @@ def lambda_handler(event, context):
         image_time = str(labeled["DateTimeOriginal"]).split(' ')[1]
         labeled['image_time'] = image_time
         labeled['image_date'] = image_date
-        geotags = get_geotagging(exif)
-        geotags = get_geotagging(exif)
-        if geotags == 'UNK':
-            pass
-        else:
-            lat_lon = get_coordinates(geotags)
-            labeled['lat_lon'] = f"{lat_lon[0]}, {lat_lon[1]}"
-            labeled['altitude'] = (geotags['GPSAltitude'])
+        gonet_software_version = str(labeled["Software"]).split(' ')[1]
+        print(gonet_software_version)
+        has_artist = float(gonet_software_version) >= 21 
+        #* V21 software added Artist Tag
+        if has_artist:
+            print("Has Artist EXIF Tag")
+            artist_lat = str(labeled["Artist"]).split(' ')[8]
+            artist_lon = str(labeled["Artist"]).split(' ')[10].rstrip(',')
+            labeled['lat_lon'] = f"{artist_lat} {artist_lon}"
+            labeled['altitude'] = str(labeled["Artist"]).split(' ')[12]
+        else: # decode the GEOtags to get Alt and Lat-Lon
+            geotags = get_geotagging(exif)
+            if geotags == 'UNK':
+                pass
+            else:
+                lat_lon = get_coordinates(geotags)
+                labeled['lat_lon'] = f"{lat_lon[0]}, {lat_lon[1]}"
+                labeled['altitude'] = (geotags['GPSAltitude'])
 
-    # format GONet Custom EXIF
+    #* format GONet Custom EXIF
     gonet_camera_name = str(labeled["Software"]).split(' ')[0]
     gonet_software_version = str(labeled["Software"]).split(' ')[1]
     gonet_white_balance = str(labeled["Software"]).split(
         ' ')[3] + str(labeled["Software"]).split(' ')[4]
-
-    print(f"Geotags:: {geotags}")
+    has_artist = float(gonet_software_version) > 20
 
     try:
         ddb_dict = {}
@@ -183,13 +207,13 @@ def lambda_handler(event, context):
                 ddb_dict[key] = str(val)
         ddb_dict['item_added'] = td
         table.put_item(Item=ddb_dict)
-        print(f"Added Item to DynamoDB Table - {Envs.ddb_table} :: {ddb_dict}")
+        # print(f"Added Item to DynamoDB Table - {Envs.ddb_table} :: {ddb_dict}")
     except botocore.exceptions.ClientError as e:
         print(f"ERROR! - {e}")
-    # print(f"Added Item to DynamoDB Table - {Envs.ddb_table} :: {ddb_dict}")
-    ## cleanup /tmp
+    print(f"EXIF data added to DDB Table: {Envs.ddb_table}")
+    #* cleanup /tmp
     print("Cleaning up /tmp")
     os.remove(source_image_tmp)
     os.remove(f"/tmp/{tiff_filename}")
     os.remove(f"/tmp/{jpeg_filename}")
-    print('## DONE')
+    print('#### DONE ####')
